[Steam] No regent reelection for Anbennar custom Republics
Regents can no longer be reelected in Technocracy and Racial Equality elections.
It is to prevent (mostly for technocracy) a bug where an immortal consort can stay in power semi-indefinitely.
Currently, the events that kill rulers/heirs/consorts don't work at all for regencies, so it's a bandaid fix until the bigger stuff is fixed.